requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flask-jwt-extended==4.6.0
flasgger==0.9.7.1
python-dotenv==1.0.1
Werkzeug==3.0.3
pytest==8.2.1
psycopg2-binary==2.9.9